feat(app/storage/mail.ts): creation and implementation a new storage for mail
