+++ practical-test/app/components/AppForm.vue
@@ -1,47 +1,18 @@
 <script setup>
-const form = ref({ name: '', email: '', message: '' })
-const loading = ref(false)
-const success = ref(false)
-const error = ref(false)
-const submit = async () => {
-  loading.value = true
-  try {
-    const res = await fetch('https://formspree.io/f/xnjglnvv', {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(form.value),
-    })
-    success.value = res.ok
-
-    if (res.ok) {
-      form.value = { name: '', email: '', message: '' }
-      success.value = true
-      setTimeout(() => {
-        success.value = false
-      }, 3000)
-    }
-  } catch {
-    error.value = true
-    setTimeout(() => {
-      error.value = false
-    }, 3000)
-  } finally {
-    loading.value = false
-  }
-}
+const form = useForm()
 </script>
 <template>
   <section
     class="w-full border border-pink-100 rounded-3xl p-10 bg-card-bg shadow"
   >
     <form
-      @submit.prevent="submit"
+      @submit.prevent="form.submit"
       class="flex flex-col md:grid md:grid-cols-2 gap-7 text-left"
     >
       <div class="flex flex-col">
         <h3 class="font-semibold pb-3 text-sm text-project-text">NAME</h3>
         <input
-          v-model="form.name"
+          v-model="form.form.name"
           placeholder="John Doe"
           required
           class="p-3 rounded-lg w-full placeholder:text-form-text border focus:border-hero-last focus:outline-none focus:border-2"
@@ -51,7 +22,7 @@ const submit = async () => {
       <div class="flex flex-col">
         <h3 class="font-semibold pb-3 text-sm text-project-text">EMAIL</h3>
         <input
-          v-model="form.email"
+          v-model="form.form.email"
           type="email"
           placeholder="john@example.com"
           required
@@ -62,7 +33,7 @@ const submit = async () => {
       <div class="md:col-span-2 flex flex-col">
         <h3 class="font-semibold pb-3 text-sm text-project-text">MESSAGE</h3>
         <textarea
-          v-model="form.message"
+          v-model="form.form.message"
           id="message"
           placeholder="Your idea..."
           required
@@ -72,14 +43,14 @@ const submit = async () => {
 
       <AppButton
         type="submit"
-        :disabled="loading"
+        :disabled="form.loading"
         styles="w-full rounded-lg bg-purple-500 md:col-span-2 p-4"
       >
-        {{ loading ? 'Sending...' : 'Send' }}
+        {{ form.loading ? 'Sending...' : 'Send' }}
       </AppButton>
 
-      <p v-if="success" class="text-purple-900">¡Message sent!</p>
-      <p v-if="error" class="text-red-800">Error sending</p>
+      <p v-if="form.success" class="text-purple-900">¡Message sent!</p>
+      <p v-if="form.error" class="text-red-800">Error sending</p>
     </form>
   </section>
 </template>
