+++ practical-test/nuxt.config.ts
@@ -4,7 +4,7 @@ import svgLoader from 'vite-svg-loader' // <-- IMPORTA EL PLUGIN
 export default defineNuxtConfig({
   compatibilityDate: '2025-07-15',
   devtools: { enabled: true },
-  modules: ['@nuxtjs/tailwindcss'],
+  modules: ['@nuxtjs/tailwindcss', '@pinia/nuxt'],
   app: {
     head: {
       link: [
